Create a build script to generate shell completion
